+++ iUAEParents/MultiPeerConnectivityController.h
@@ -31,10 +31,10 @@ static NSString *const kiCadePad = @"iCadePad";
 
 - (void)configure: (MainEmulationViewController *) mainEmuViewCtrl ;
 + (MultiPeerConnectivityController *)getinstance;
-- (void)handleinputdirections:(TouchStickDPadState)hat_state buttontoreleasevertical:(int)buttontoreleasevertical buttontoreleasehorizontal: (int)buttontoreleasehorizontal deviceid:(NSString *)dID;
-- (int)handleinputbuttons:(int)buttonid buttonstate:(int)buttonstate deviceid:(NSString *)dID;
-- (void)handleinputdirections:(TouchStickDPadState)hat_state buttontoreleasevertical:(int)buttontoreleasevertical buttontoreleasehorizontal: (int)buttontoreleasehorizontal deviceid: (NSString *)dID;
-- (int) dpadstatetojoypadkey:(NSString *)direction hatstate:(TouchStickDPadState)hat_state;
+- (void)handleinputdirections:(TouchStickDPadState)hat_state buttontoreleasevertical:(NSInteger)buttontoreleasevertical buttontoreleasehorizontal: (NSInteger)buttontoreleasehorizontal deviceid:(NSString *)dID;
+- (NSInteger)handleinputbuttons:(NSInteger)buttonid buttonstate:(NSInteger)buttonstate deviceid:(NSString *)dID;
+- (void)handleinputdirections:(TouchStickDPadState)hat_state buttontoreleasevertical:(NSInteger)buttontoreleasevertical buttontoreleasehorizontal: (NSInteger)buttontoreleasehorizontal deviceid: (NSString *)dID;
+- (NSInteger) dpadstatetojoypadkey:(NSString *)direction hatstate:(TouchStickDPadState)hat_state;
 - (void)enableControllerMode;
 - (void)controllerDisconnected:(NSString *)dID;
