+++ uae4all_gp2x_0.7.2a/src/drawing.cpp
@@ -523,14 +523,14 @@ static __inline__ void fill_line (void)
     dbgf("fill_line -> nints=%i, nrem=%i, val=%i\n",nints,nrem,val);
 #endif
     for (; nints > 0; nints -= 8, start += 8) {
-		*start = val;
-		*(start+1) = val;
-		*(start+2) = val;
-		*(start+3) = val;
-		*(start+4) = val;
-		*(start+5) = val;
-		*(start+6) = val;
-		*(start+7) = val;
+		*start = (int)val;
+		*(start+1) = (int)val;
+		*(start+2) = (int)val;
+		*(start+3) = (int)val;
+		*(start+4) = (int)val;
+		*(start+5) = (int)val;
+		*(start+6) = (int)val;
+		*(start+7) = (int)val;
     }
 	
 	/*gno: no reminders
@@ -1844,7 +1844,7 @@ static _INLINE_ void do_color_changes (line_draw_func worker_border, line_draw_f
 	
     for (i = dip_for_drawing->first_color_change; i <= dip_for_drawing->last_color_change; i++) {
 		int regno = curr_color_changes[i].regno;
-		unsigned int value = curr_color_changes[i].value;
+		unsigned int value = (unsigned int)curr_color_changes[i].value;
 		int nextpos, nextpos_in_range;
 		if (i == dip_for_drawing->last_color_change)
 			nextpos = max_diwlastword;
@@ -2121,7 +2121,7 @@ static _INLINE_ void draw_status_line (int line)
                     break;
             }
         }
-        c = xcolors[on ? on_rgb : off_rgb];
+        c = (int)xcolors[on ? on_rgb : off_rgb];
         
         for (j = 0; j < TD_LED_WIDTH; j++)
             putpixel (x + j, c);
@@ -2171,7 +2171,7 @@ static void fps_counter_upd(void)
 	gettimeofday(&tv, 0);
 	if (tv.tv_sec != thissec)
 	{
-		thissec = tv.tv_sec;
+		thissec = (int)tv.tv_sec;
 		fps_counter = fcount;
 		fcount = 0;
 		fps_counter_changed = 1;
